Random word endpoint (#29)
* feat(random_word): :sparkles: Define word, phoneme, wordphonemelink models and test database creation with these models

* style(random_word): :art: Format with ruff

* feat(random_word): :construction: Draft random word endpoint without phonemes

* test(test): :white_check_mark: Add pytest infrastructure for testing within seeded or unseeded database

Default seeding is "software" and "hardware". Custom seedings can be defined by using @pytest.mark.parametrize and specifying specific test_seed_data

* test(random_word): :test_tube: Add failing random word tests within new pytest infrastructure

Fails due to non-unique word_id-phoneme_id pairs

* fix(Models): :bug: Use index as part of primary key for WordPhonemeLink

Needed to identify unique instances of occurrences of a phoneme within a word. For example, "r" occurs twice in "hardware"

* test(random_word): :white_check_mark: Test phonemes returned alongside random word

* refactor(Models): :recycle: Suggest refactoring of Word to remove unnecessary phonemes relationship

We always want phonemes to be ordered by their index in a word, having an unordered list of phonemes for a given word is not useful

* refactor(Schemas): :recycle: Extract phoneme schema as it will be later shared in other endpoints (e.g. recording)

* refactor(random_word): :bulb: Remove comments for PR

---------

Co-authored-by: Jack Nguyen <trungnb2210@gmail.com>